minor fix for factories and testing

+++ sail/csrc/src/factories.h
@@ -33,10 +33,10 @@ namespace random {  // probably want to refactor factories to be in their own
                     // namespace but rolling with this for now
 
 // need to be able to instantiate random tensors
-Tensor uniform(TensorShape size, Dtype dt, int min = 0, int max = 1);
-Tensor uniform_like(Tensor tensor, int min = 0, int max = 1);
-Tensor normal(TensorShape size, Dtype dt, int mean = 0, int std = 1);
-Tensor normal_like(Tensor tensor, int mean = 0, int std = 1);
+Tensor uniform(TensorShape size, Dtype dt, double min = 0, double max = 1);
+Tensor uniform_like(Tensor tensor, double min = 0, double max = 1);
+Tensor normal(TensorShape size, Dtype dt, double mean = 0, double std = 1);
+Tensor normal_like(Tensor tensor, double mean = 0, double std = 1);
 }  // namespace random
 
 }  // namespace sail
